[수정] `onChange` 제거 및 학과검색 데이터모델 생성
