+++ package-kuring/Sources/UIKit/OnboardingUI/Common/StringSet.swift
@@ -1,0 +1,21 @@
+//
+// Copyright (c) 2024 쿠링
+// See the 'License.txt' file for licensing information.
+//
+
+import Foundation
+
+enum StringSet: String {
+    case title = "우리 대학 공지\n쿠링이 알려줄게!"
+    case button_goSubscrbing = "공지 알림 설정하러 가기"
+    
+    case title_select = "전공학과를\n설정해주세요"
+    case description_select = "학과 공지 알림을 받아볼 수 있도록\n전공 학과를 설정해 주세요"
+    case button_complete = "학과 설정 완료"
+    case button_again = "다시 설정하기"
+    case button_start = "쿠링 시작하기"
+    case title_confirm = "내 전공학과로 설정할까요?"
+    case description_confirm = "혹시 다전공/부전공을 하고 계신가요?\n쿠링을 시작하면 더 많은 전공을 추가할 수 있어요!"
+    case title_complete = "학과 설정이 완료되었어요!"
+    case description_complete = "이제 쿠링과 함께\n즐겁고 편리한 캠퍼스를 누려봐요!"
+}
