+++ package-kuring/Sources/UIKit/OnboardingUI/MyDepartmentSelector/MyDepartmentSelector.Step.swift
@@ -1,0 +1,16 @@
+//
+// Copyright (c) 2024 쿠링
+// See the 'License.txt' file for licensing information.
+//
+
+import Foundation
+
+extension MyDepartmentSelector {
+    enum Step: Identifiable {
+        case searchDepartment
+        case selectDepartment
+        case addedDepartment
+        
+        var id: Self { self }
+    }
+}
