--- package-kuring/Sources/UIKit/OnboardingUI/OnboardingView.StringSet.swift
@@ -1,13 +1,0 @@
-//
-// Copyright (c) 2024 쿠링
-// See the 'License.txt' file for licensing information.
-//
-
-import Foundation
-
-extension OnboardingView {
-    enum StringSet: String {
-        case title = "우리 대학 공지\n쿠링이 알려줄게!"
-        case button_goSubscrbing = "공지 알림 설정하러 가기"
-    }
-}
